chore: cleanup map elements
