+++ src/app/components/map/marker/BusMarker.tsx
@@ -1,10 +1,11 @@
 import useAppStore from '@data/state/app_state';
 import { Bus, Route } from '@data/types';
+import { getLighterColor } from '@data/utils/utils';
+import { MaterialCommunityIcons } from '@expo/vector-icons';
 import React, { memo } from 'react';
-import { Platform } from 'react-native';
+import { Platform, View } from 'react-native';
 import { Marker } from 'react-native-maps';
-import BusCallout from '../BusCallout';
-import BusMapIcon from '../mapIcons/BusMapIcon';
+import BusCallout from '../callout/BusCallout';
 
 interface Props {
   bus: Bus;
@@ -25,6 +26,18 @@ const BusMarker: React.FC<Props> = ({ bus, route }) => {
       setSelectedDirection(bus.direction);
     }
   };
+
+  const getRotationProp = (bearing: number | undefined) => {
+    return [
+      {
+        rotate:
+          bearing !== undefined ? `${Math.round(bearing) - 135}deg` : '0deg',
+      },
+    ];
+  };
+
+  const active = bus.direction.id === selectedDirection?.id;
+  const borderColor = active ? getLighterColor(route.tintColor) : undefined;
 
   return (
     <Marker
@@ -49,12 +62,30 @@ const BusMarker: React.FC<Props> = ({ bus, route }) => {
       onPress={() => busDefaultDirection()}
     >
       {/* Bus Icon on Map*/}
-      <BusMapIcon
-        bus={bus}
-        route={route}
-        selectedDirection={selectedDirection}
-      />
-
+      <View
+        style={{
+          alignItems: 'center',
+          justifyContent: 'center',
+          width: 30,
+          height: 30,
+          borderTopLeftRadius: 15,
+          borderTopRightRadius: 15,
+          borderBottomLeftRadius: 15,
+          backgroundColor: active ? route.tintColor : route.tintColor + '70',
+          borderColor: borderColor,
+          borderWidth: borderColor ? 2 : 0,
+          transform: getRotationProp(bus.heading),
+          zIndex: active ? 800 : 500,
+          elevation: active ? 800 : 500,
+        }}
+      >
+        <MaterialCommunityIcons
+          name="bus"
+          size={18}
+          color={active ? 'white' : '#ffffffcc'}
+          style={{ transform: getRotationProp(-bus.heading - 90) }}
+        />
+      </View>
       <BusCallout route={route} bus={bus} />
     </Marker>
   );
